requirements.txt
flask>=3.0
gunicorn>=21.2

cryptography>=42.0

boto3>=1.34
requests>=2.31

openai>=1.30
chromadb>=0.5

# LlamaIndex RAG pipeline
llama-index-core>=0.11
llama-index-readers-file>=0.4
llama-index-vector-stores-chroma>=0.4
llama-index-embeddings-openai>=0.3
llama-index-llms-openai>=0.3
pymupdf>=1.24

prowler>=5.18.3
fpdf2>=2.8

pandas>=2.1

streamlit>=1.35

vercel>=0.5.1